requirements.txt
python-dotenv==0.20.0
boto3==1.24.59
ScraperFC==2.0.2
numpy
matplotlib